Move Cat function to chisel3 package from chisel3.util
